third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/fast/frames/iframe-onload-and-domnodeinserted-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
This test ensures that any tree mutation in the load event handler cannot harm the tree consistency.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS loadEventFired is true
PASS unless crash.
PASS successfullyParsed is true

TEST COMPLETE
